+++ problems/0052/solution.jl
@@ -1,7 +1,7 @@
 using BenchmarkTools
 BenchmarkTools.DEFAULT_PARAMETERS.samples = 100
 
-function compute(n::Int64)::Int64
+function compute(n::Int)::Int
     i = 1
     while true
         for j ∈ 10 ^ i:10 ^ (i + 1) ÷ n - 1
